#0: commonize some functions
