+++ tests/tt_metal/tt_metal/test_kernels/dataflow/unit_tests/erisc/ethernet_write_worker_latency_ubench_common.hpp
@@ -1,0 +1,37 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include <cstdint>
+#include <array>
+#include "eth_l1_address_map.h"
+#include "dataflow_api.h"
+#include "ethernet/dataflow_api.h"
+#include "debug/assert.h"
+#include "debug/dprint.h"
+
+// #define ENABLE_DEBUG 1
+
+struct eth_buffer_slot_sync_t {
+    volatile uint32_t bytes_sent;
+    volatile uint32_t receiver_ack;
+    volatile uint32_t src_id;
+
+    uint32_t reserved_2;
+};
+
+FORCE_INLINE void eth_setup_handshake(std::uint32_t handshake_register_address, bool is_sender) {
+    if (is_sender) {
+        eth_send_bytes(handshake_register_address, handshake_register_address, 16);
+        eth_wait_for_receiver_done();
+    } else {
+        eth_wait_for_bytes(16);
+        eth_receiver_channel_done(0);
+    }
+}
+
+FORCE_INLINE void switch_context_if_debug() {
+#if ENABLE_DEBUG
+    internal_::risc_context_switch();
+#endif
+}
